:recycle: 一覧ページの構造をロジックと表示部分に分割
