+++ src/client/containers/contributes/List/style.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/colors';
-@import '../../styles/font-size';
+@import '../../../styles/colors';
+@import '../../../styles/font-size';
 
 .floatButtonContainer {
   position: fixed;
