+++ src/pages/index.tsx
@@ -1,3 +1,3 @@
-import Home from "@containers/contributes/Index";
+import Home from "@containers/contributes/List";
 
 export default Home;
